feat: 선택에 대한 정의 Selectable 프로토콜로 분리
